pyproject.toml
[tool.poetry]
name = "django_restframework"
version = "0.1.0"
description = ""
authors = ["oner <caimbebr@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.10"
Django = "^4.1"
djangorestframework = "^3.13.1"
Markdown = "^3.4.1"
blue = "^0.9.1"
isort = "^5.10.1"
pylint = "^2.14.5"
prospector = "^1.7.7"
pip-audit = "^2.4.3"
pylint-django = "^2.5.3"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
